Scala.par.map shows a behaviour where the job starts with fill CPU utilisation but limbs home on 1 thread. It may be allocating tasks to threads upfront so preforms badly if the jobs are not evenly sized.
Old school Java ThreadPool does max out the CPU right to the very last tasks.
GB 1901001 in real	5m14.347s
